+++ game/data/resources.ts
@@ -9,18 +9,12 @@ export type ResourceData = {
   default_value?: DecimalSource;
 };
 
-export type ResourceNames = "wood" | "stone" | "food" | "wooden spear";
+export type ResourceNames = "wood" | "stone" | "food";
 
 export const Resources: { [key in ResourceNames]: ResourceData } = {
   wood: {},
   stone: {},
   food: {
     default_value: 100,
-  },
-  "wooden spear": {
-    cost: {
-      wood: 10,
-      stone: 10,
-    },
   },
 };
