+++ src/pages/StageOnePage.vue
