+++ src/pages/Statistics.vue
